+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added bandits, track you if you have a cabin (because they see the smoke)
 - Added alert for when the cell the player is on is empty
 - Added **Scythe** item that allows you to gather food more quickly from fields
+- Added **Cooking** technology to automatically eat when hungry
 
 ### Changed
 - Made the starting cell always populated wiht some resources